requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
demucs==4.0.1
opencv-python==4.8.1.78
moviepy==1.0.3
Pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
openai==1.3.7
numpy==1.24.3
scipy==1.11.4
matplotlib==3.7.2
ffmpeg-python==0.2.0
mutagen==1.47.0
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
sqlalchemy
psycopg2-binary
alembic
python-multipart
